Fixed a couple of signed comparisons

+++ src/sdf/idata_file.cc
@@ -23,7 +23,7 @@
 using namespace sdf;
 
 static std::streamsize safe_streamsize(size_t size) {
-  if (size > std::numeric_limits<std::streamsize>::max()) {
+  if (size > static_cast<size_t>(std::numeric_limits<std::streamsize>::max())) {
     throw std::overflow_error("Data record too large for std::istream");
   }
   return static_cast<std::streamsize>(size);
